feat: enhance game popup actions with confirmation callbacks
