+++ service/app/src/entities/game/model/useGamePopupActions.tsx
@@ -4,62 +4,31 @@ import {
   Dialog,
   DialogBody,
   DialogButton,
-  DialogClose,
   DialogContent,
   DialogFooter,
   DialogHeader,
 } from "@shared/design/src/components/dialog"
 import { overlay } from "overlay-kit"
 
-import { saveGame } from "../utils/gameSave"
-import { useGameCreationContext } from "./state/create/gameCreationContext"
-import { selectors } from "./state/create/selectors"
-
 export function useGamePopupActions() {
-  const { state, actions } = useGameCreationContext()
-
-  const showSaveConfirm = () => {
+  const showSaveConfirm = (onConfirm: () => void) => {
     overlay.open(({ isOpen, close }) => {
       return (
         <Dialog open={isOpen} onOpenChange={() => close()}>
           <DialogContent>
             <DialogHeader>게임을 저장하시겠습니까?</DialogHeader>
             <DialogFooter variant="onlyTitle">
-              <DialogClose asChild>
-                <DialogButton.Secondary>아니요</DialogButton.Secondary>
-              </DialogClose>
-              <DialogClose asChild>
-                <DialogButton.Primary
-                  onClick={async () => {
-                    try {
-                      actions.saveGameStart()
-
-                      const cleanedQuestions = selectors.cleanedQuestions(state)
-
-                      const result = await saveGame({
-                        ...state,
-                        questions: cleanedQuestions,
-                      })
-
-                      if (result.success) {
-                        actions.saveGameSuccess()
-                      } else {
-                        actions.saveGameError(
-                          result.error || "저장에 실패했습니다.",
-                        )
-                        close()
-                        showSaveError()
-                      }
-                    } catch (error) {
-                      actions.saveGameError("알 수 없는 오류가 발생했습니다.")
-                      close()
-                      showSaveError()
-                    }
-                  }}
-                >
-                  네
-                </DialogButton.Primary>
-              </DialogClose>
+              <DialogButton.Secondary onClick={() => close()}>
+                아니요
+              </DialogButton.Secondary>
+              <DialogButton.Primary
+                onClick={() => {
+                  onConfirm()
+                  close()
+                }}
+              >
+                네
+              </DialogButton.Primary>
             </DialogFooter>
           </DialogContent>
         </Dialog>
@@ -76,9 +45,9 @@ export function useGamePopupActions() {
               JPG, JPEG, PNG 형식만 가능하며, 최대 2MB까지 업로드할 수 있습니다.
             </DialogBody>
             <DialogFooter variant="onlyBody">
-              <DialogClose asChild>
-                <DialogButton.Secondary>닫기</DialogButton.Secondary>
-              </DialogClose>
+              <DialogButton.Secondary onClick={() => close()}>
+                닫기
+              </DialogButton.Secondary>
             </DialogFooter>
           </DialogContent>
         </Dialog>
@@ -86,7 +55,7 @@ export function useGamePopupActions() {
     })
   }
 
-  const showLibraryRegister = () => {
+  const showLibraryRegister = (onConfirm: () => void) => {
     overlay.open(({ isOpen, close }) => {
       return (
         <Dialog open={isOpen} onOpenChange={() => close()}>
@@ -96,12 +65,17 @@ export function useGamePopupActions() {
             </DialogHeader>
             <DialogBody>저장하지 않으면 모든 변경사항이 사라집니다.</DialogBody>
             <DialogFooter variant="title">
-              <DialogClose asChild>
-                <DialogButton.Secondary>아니요</DialogButton.Secondary>
-              </DialogClose>
-              <DialogClose asChild>
-                <DialogButton.Primary>네</DialogButton.Primary>
-              </DialogClose>
+              <DialogButton.Secondary onClick={() => close()}>
+                아니요
+              </DialogButton.Secondary>
+              <DialogButton.Primary
+                onClick={() => {
+                  onConfirm()
+                  close()
+                }}
+              >
+                네
+              </DialogButton.Primary>
             </DialogFooter>
           </DialogContent>
         </Dialog>
@@ -120,9 +94,9 @@ export function useGamePopupActions() {
               다시 시도해 주세요.
             </DialogBody>
             <DialogFooter variant="onlyBody">
-              <DialogClose asChild>
-                <DialogButton.Secondary>닫기</DialogButton.Secondary>
-              </DialogClose>
+              <DialogButton.Secondary onClick={() => close()}>
+                닫기
+              </DialogButton.Secondary>
             </DialogFooter>
           </DialogContent>
         </Dialog>
@@ -140,9 +114,9 @@ export function useGamePopupActions() {
               모든 필수 항목을 작성한 후 다시 저장해 주세요.
             </DialogBody>
             <DialogFooter variant="onlyBody">
-              <DialogClose asChild>
-                <DialogButton.Secondary>닫기</DialogButton.Secondary>
-              </DialogClose>
+              <DialogButton.Secondary onClick={() => close()}>
+                닫기
+              </DialogButton.Secondary>
             </DialogFooter>
           </DialogContent>
         </Dialog>
