+++ service/app/src/entities/game/ui/interactions/registerButton.tsx
@@ -1,13 +1,18 @@
 import { SecondaryPlainIconButton } from "@shared/design/src/components/button"
 import { Cross } from "@shared/design/src/icons"
+import { useRouter } from "next/navigation"
 
 import { useGamePopupActions } from "../../model/useGamePopupActions"
 
 export function RegisterButton() {
+  const router = useRouter()
   const { showLibraryRegister } = useGamePopupActions()
 
   const handleRegister = () => {
-    showLibraryRegister()
+    showLibraryRegister(() => {
+      // 게임 생성 페이지에서 나가기
+      router.push("/dashboard")
+    })
   }
 
   return (
