+++ service/app/src/entities/game/ui/interactions/saveButton.tsx
@@ -2,11 +2,13 @@ import { PrimaryBoxButton } from "@shared/design/src/components/button"
 
 import { validateQuestion } from "../../model"
 import { useGameCreationContext } from "../../model/state/create/gameCreationContext"
+import { selectors } from "../../model/state/create/selectors"
 import { useGamePopupActions } from "../../model/useGamePopupActions"
+import { saveGame } from "../../utils/gameSave"
 
 export function SaveButton() {
-  const { state } = useGameCreationContext()
-  const { showSaveConfirm, showValidationError } = useGamePopupActions()
+  const { state, actions } = useGameCreationContext()
+  const { showSaveConfirm, showValidationError, showSaveError } = useGamePopupActions()
 
   const handleSave = () => {
     const hasValidationError = state.questions.some(
@@ -18,7 +20,30 @@ export function SaveButton() {
       return
     }
 
-    showSaveConfirm()
+    showSaveConfirm(async () => {
+      try {
+        actions.saveGameStart()
+
+        const cleanedQuestions = selectors.cleanedQuestions(state)
+
+        const result = await saveGame({
+          ...state,
+          questions: cleanedQuestions,
+        })
+
+        if (result.success) {
+          actions.saveGameSuccess()
+        } else {
+          actions.saveGameError(
+            result.error || "저장에 실패했습니다.",
+          )
+          showSaveError()
+        }
+      } catch (error) {
+        actions.saveGameError("알 수 없는 오류가 발생했습니다.")
+        showSaveError()
+      }
+    })
   }
 
   return (
